rename Company.tsx -> CompanyListItem.tsx
