+++ app/CompanyListItem.tsx
